+++ client/package.json
@@ -11,13 +11,14 @@
     "clsx": "^1.0.4",
     "formik": "^2.1.1",
     "material-table": "^1.55.0",
-    "node-sass": "^4.13.0",
     "react": "^16.12.0",
     "react-bootstrap": "^1.0.0-beta.16",
     "react-dom": "^16.12.0",
+    "react-is": "^16.12.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.3.0",
     "react-table": "^7.0.0-rc.15",
+    "styled-components": "^5.0.0",
     "yup": "^0.28.0"
   },
   "scripts": {
