--- client/public/assets/css/loading-bar.css
@@ -1,18 +1,0 @@
-.ldBar {
-  position: relative;
-}
-.ldBar.label-center > .ldBar-label {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-  text-shadow: 0 0 3px #fff;
-}
-.ldBar-label:after {
-  content: "%";
-  display: inline;
-}
-.ldBar.no-percent .ldBar-label:after {
-  content: "";
-}
