+++ client/public/index.html
@@ -24,8 +24,6 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
-    <link rel="stylesheet" type="text/css" href="./assets/css/loading-bar.css" />
-    <script type="text/javascript" src="./assets/js/loading-bar.js"></script>
     <title>React App</title>
   </head>
   <body>
