+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import { Footer, SideNav } from 'components'
+import { Footer } from 'components'
 import {
   About,
   User,
@@ -8,28 +8,25 @@ import {
   Signup,
   Trend,
   Watchlist,
-  PageNotFound
+  PageNotFound,
+  Layout
 } from './pages'
 import useMediaQuery from '@material-ui/core/useMediaQuery'
 import { ThemeProvider, createMuiTheme } from '@material-ui/core/styles'
-import './App.scss'
+import { DarkMode } from 'components/_common/DarkMode'
 
 export default function App() {
-  const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)')
-  
-  const theme = React.useMemo(
-    () =>
-      createMuiTheme({
-        palette: {
-          type: prefersDarkMode ? 'dark' : 'light'
-        }
-      }),
-    [prefersDarkMode]
+  const theme = React.useMemo(() =>
+    createMuiTheme({
+      palette: {
+        type: 'dark'
+      }
+    })
   )
 
   return (
     <ThemeProvider theme={theme}>
-      <SideNav>
+      <Layout>
         <Router>
           <Switch>
             <Route exact path='/' component={Signup} />
@@ -41,7 +38,7 @@ export default function App() {
             <Route component={PageNotFound} />
           </Switch>
         </Router>
-      </SideNav>
+      </Layout>
       <Footer />
     </ThemeProvider>
   )
