--- client/src/components/Button/styles.scss
@@ -1,8 +1,0 @@
-.material-icons {
-  position: relative;
-  padding-top: 5%;
-}
-
-.material-icons.md-20 {
-  font-size: 20px;
-}
