+++ client/src/components/Card/_Template/index.jsx
@@ -1,15 +1,13 @@
 import React from 'react'
-import CardColumns from 'react-bootstrap/CardColumns'
 import { TrendCard } from 'components'
-import './styles.scss'
 
 export const Columns = props => {
   const { tests, onDelete } = props
   return (
-    <CardColumns className='cardColumns'>
+    <div>
       {tests.map(test => (
         <TrendCard test={test} onDelete={onDelete} />
       ))}
-    </CardColumns>
+    </div>
   )
 }
