--- client/src/components/Card/styles.scss
@@ -1,10 +1,0 @@
-.card {
-  text-align: center;
-  width: 100%;
-  height: 100%;
-  padding: 10px;
-}
-
-.close{
-  float: right;
-}
