--- client/src/components/CircleBar/styles.scss
@@ -1,21 +1,0 @@
-.ldBar {
-  margin: auto;
-}
-
-.ldBar path.mainline {
-  stroke-width: 10;
-  stroke: #09f;
-  stroke-linecap: round;
-}
-
-.ldBar path.baseline {
-  stroke-width: 14;
-  stroke: #f1f2f3;
-}
-
-.ldBar-label {
-  color: #09f;
-  font-family: 'varela round';
-  font-size: 1em;
-  font-weight: 900;
-}
