--- client/src/components/Columns/styles.scss
@@ -1,4 +1,0 @@
-.cardColumns {
-	width: 900px;
-	margin: 0 auto;
-}
