--- client/src/components/Footer/styles.scss
@@ -1,5 +1,0 @@
-.footer {
-  position: absolute;
-  bottom: 0;
-  overflow: auto;
-}
