+++ client/src/components/Menu/Notification_Menu/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import Badge from '@material-ui/core/Badge'
+import IconButton from '@material-ui/core/IconButton'
+import NotificationsIcon from '@material-ui/icons/Notifications'
+
+export const NotificationMenu = () => {
+  return (
+    <IconButton aria-label='show 17 new notifications' color='inherit'>
+      <Badge badgeContent={17} color='secondary'>
+        <NotificationsIcon />
+      </Badge>
+    </IconButton>
+  )
+}
