+++ client/src/components/Menu/index.js
@@ -1,0 +1,5 @@
+export * from './Account_Menu'
+export * from './AppBar_Menu'
+export * from './Drawer_Menu'
+export * from './Filter_Menu'
+export * from './Notification_Menu'
