+++ client/src/components/index.js
@@ -1,10 +1,4 @@
 export * from './Card'
-export * from './Columns'
 export * from './Footer'
-export * from './Navbar'
 export * from './Form'
-export * from './Button'
-export * from './CircleBar'
 export * from './Table'
-export * from './FilterMenu'
-export * from './SideNav'
