+++ client/src/index.js
@@ -4,7 +4,5 @@ import App from "./App";
 import registerServiceWorker from "./registerServiceWorker";
 import 'bootstrap/dist/css/bootstrap.min.css'
 
-import "./index.scss";
-
 ReactDOM.render(<App />, document.getElementById("root"));
 registerServiceWorker();
