--- client/src/index.scss
@@ -1,7 +1,0 @@
-html,
-body {
-  margin: 0;
-  height: 100vh;
-  padding: 0;
-  font-family: sans-serif;
-}
