+++ client/src/pages/_Template/index.jsx
@@ -1,0 +1,55 @@
+import React, { children } from 'react'
+import clsx from 'clsx'
+import { makeStyles, useTheme } from '@material-ui/core/styles'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import Paper from '@material-ui/core/Paper'
+import { DrawerMenu, AppBarMenu } from 'components/Menu'
+
+const drawerWidth = 200
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: 'flex',
+    backgroundColor: theme.palette.background.default,
+    color: theme.palette.text.primary
+  },
+  toolbar: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'flex-end',
+    padding: theme.spacing(0, 1),
+    ...theme.mixins.toolbar
+  },
+  content: {
+    flexGrow: 1,
+    padding: theme.spacing(3),
+    height: 600
+  }
+}))
+
+export function Layout({ children }) {
+  const classes = useStyles()
+  const theme = useTheme()
+  const [open, setOpen] = React.useState(false)
+
+  const handleDrawerOpen = () => {
+    setOpen(true)
+  }
+
+  const handleDrawerClose = () => {
+    setOpen(false)
+  }
+
+  return (
+    <div className={classes.root}>
+      <CssBaseline />
+      <AppBarMenu />
+      <DrawerMenu />
+
+      <main className={classes.content}>
+        <div className={classes.toolbar} />
+        <Paper>{children}</Paper>
+      </main>
+    </div>
+  )
+}
