+++ client/src/pages/index.js
@@ -5,3 +5,4 @@ export * from './Watchlist'
 export * from './User'
 export * from './About'
 export * from './PageNotFound'
+export * from './_Template'
