+++ package.json
@@ -21,8 +21,13 @@
   },
   "dependencies": {
     "axios": "^0.19.1",
+    "bcryptjs": "^2.4.3",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "if-env": "^1.0.4",
-    "mongoose": "^5.8.6"
+    "jsonwebtoken": "^8.5.1",
+    "jwt-decode": "^2.2.0",
+    "mongoose": "^5.8.6",
+    "react-toastify": "^5.5.0"
   }
 }
